language_processing/requirements.txt
python-dotenv==1.0.1
qdrant-client==1.9.0
langchain==0.1.16
langchain-openai==0.1.3
pymongo==4.6.3
flask==3.0.3
nltk==3.8.1
gunicorn==22.0.0
requests